git/bitbucket/create-pull-request/requirements.txt
oauthlib==3.1.0
python-dotenv
requests
requests_oauthlib
